+++ packages/helpers/src/builders/components/MessageActionRowBuilder.ts
@@ -1,10 +1,10 @@
-import type { DiscordActionRow, MessageComponentTypes } from '@biscuitland/api-types';
-import type { ComponentBuilder } from '../../utils/util';
+import { DiscordActionRow, MessageComponentTypes } from '@biscuitland/api-types';
+import type { ComponentBuilder } from '../../../../core/src/utils/util';
 
 export class ActionRowBuilder<T extends ComponentBuilder> {
     constructor() {
         this.components = [] as T[];
-        this.type = 1;
+        this.type = MessageComponentTypes.ActionRow;
     }
     components: T[];
     type: MessageComponentTypes.ActionRow;
